refactor: reorganize test files for better structure
- Move DateTime utility method tests from api_enhancements_test.go to datetime_test.go
  - Start/End operations (StartOfDay, EndOfDay, StartOfMonth, etc.)
  - Weekend/Weekday detection (IsWeekend, IsWeekday)
  - Quarter operations (Quarter, StartOfQuarter, EndOfQuarter)
  - ISO Week operations (ISOWeek, ISOWeekYear, ISOWeekNumber, DayOfYear)

- Create dedicated fluent_test.go for Fluent API functionality
  - FluentDuration tests (calendar arithmetic accuracy)
  - FluentDateTime tests (method chaining)
  - ChronoDuration tests (enhanced duration operations)

- Delete api_enhancements_test.go as all tests properly integrated
- All 150+ tests continue to pass with improved organization
- Tests now follow logical grouping by functionality rather than chronological addition
